move diff command to schema file
