+++ cmd/atlas/internal/cmdapi/cmdapi.go
@@ -6,11 +6,19 @@
 package cmdapi
 
 import (
+	"context"
 	"encoding/csv"
 	"errors"
 	"fmt"
+	"io"
+	"net/url"
+	"os"
+	"path/filepath"
 	"strings"
 
+	"ariga.io/atlas/sql/migrate"
+	"ariga.io/atlas/sql/schema"
+	"ariga.io/atlas/sql/sqlclient"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	"github.com/zclconf/go-cty/cty"
@@ -24,6 +32,8 @@ const (
 	flagDSN         = "dsn" // deprecated in favor of flagURL
 	flagExclude     = "exclude"
 	flagFile        = "file"
+	flagFrom        = "from"
+	flagTo          = "to"
 	flagSchema      = "schema"
 	flagURL         = "url"
 	flagVar         = "var"
@@ -209,10 +219,19 @@ func addFlagExclude(set *pflag.FlagSet, target *[]string) {
 	)
 }
 
-func addFlagURL(set *pflag.FlagSet, target *string) {
+// addFlagURL adds a URL flag. If given, args[0] override the name, args[1] the shorthand.
+func addFlagURL(set *pflag.FlagSet, target *string, args ...string) {
+	name, short := flagURL, "u"
+	switch len(args) {
+	case 2:
+		short = args[1]
+		fallthrough
+	case 1:
+		name = args[0]
+	}
 	set.StringVarP(
 		target,
-		flagURL, "u",
+		name, short,
 		"",
 		"[driver://username:password@address/dbname?param=value] select a resource using the URL format",
 	)
@@ -247,4 +266,175 @@ func maySetFlag(cmd *cobra.Command, name, envVal string) error {
 		return nil
 	}
 	return cmd.Flags().Set(name, envVal)
+}
+
+type (
+	// stateReadCloser is a migrate.StateReader with an optional io.Closer.
+	stateReadCloser struct {
+		migrate.StateReader
+		io.Closer        // optional close function
+		Schema    string // in case we work on a single schema
+	}
+	// stateReaderConfig is given to stateReader.
+	stateReaderConfig struct {
+		urls    []string          // urls to create a migrate.StateReader from
+		dev     *sqlclient.Client // dev database connection
+		schemas []string          // schemas to work on
+	}
+)
+
+// stateReader returns a migrate.StateReader that reads the state from the given urls.
+func stateReader(ctx context.Context, config *stateReaderConfig) (*stateReadCloser, error) {
+	scheme, err := selectScheme(config.urls)
+	if err != nil {
+		return nil, err
+	}
+	parsed := make([]*url.URL, len(config.urls))
+	for i, u := range config.urls {
+		parsed[i], err = url.Parse(u)
+		if err != nil {
+			return nil, err
+		}
+	}
+	switch scheme {
+	// "file" scheme is valid for both migration directory and HCL paths.
+	case "file":
+		// Replaying a migration directory or evaluating an HCL file requires a dev connection.
+		if config.dev == nil {
+			return nil, errors.New("--dev-url cannot be empty")
+		}
+		if len(config.urls) > 1 {
+			// Consider urls being HCL paths.
+			return hclStateReader(ctx, config, parsed)
+		}
+		path := filepath.Join(parsed[0].Host, parsed[0].Path)
+		fi, err := os.Stat(path)
+		if err != nil {
+			return nil, err
+		}
+		if !fi.IsDir() {
+			// If there is only one url given, and it is a file, consider it an HCL path.
+			return hclStateReader(ctx, config, parsed)
+		}
+		files, err := os.ReadDir(path)
+		if err != nil {
+			return nil, err
+		}
+		// Check all files, if we find both HCL and SQL files, abort. Otherwise, proceed accordingly.
+		var hcl, sql bool
+		for _, f := range files {
+			ext := filepath.Ext(f.Name())
+			switch {
+			case hcl && ext == ".sql", sql && ext == ".hcl":
+				return nil, fmt.Errorf("ambiguos files: %q contains both SQL and HCL files", path)
+			case ext == ".hcl":
+				hcl = true
+			case ext == ".sql":
+				sql = true
+			default:
+				// unknown extension, we don't care
+			}
+		}
+		switch {
+		case hcl:
+			return hclStateReader(ctx, config, parsed)
+		case sql:
+			dir, err := dirURL(parsed[0], false)
+			if err != nil {
+				return nil, err
+			}
+			ex, err := migrate.NewExecutor(config.dev.Driver, dir, migrate.NopRevisionReadWriter{})
+			if err != nil {
+				return nil, err
+			}
+			sr, err := ex.Replay(ctx, func() migrate.StateReader {
+				if config.dev.URL.Schema != "" {
+					return migrate.SchemaConn(config.dev, "", nil)
+				}
+				return migrate.RealmConn(config.dev, nil)
+			}())
+			if err != nil && !errors.Is(err, migrate.ErrNoPendingFiles) {
+				return nil, fmt.Errorf("replaying the migration directory: %w", err)
+			}
+			return &stateReadCloser{
+				StateReader: migrate.Realm(sr),
+				Schema:      config.dev.URL.Schema,
+			}, nil
+		default:
+			return nil, fmt.Errorf("%q contains neither SQL nor HCL files", path)
+		}
+	default:
+		// All other schemes are database (or docker) connections.
+		c, err := sqlclient.Open(ctx, config.urls[0]) // call to selectScheme already checks for len > 0
+		if err != nil {
+			return nil, err
+		}
+		var sr migrate.StateReader
+		switch c.URL.Schema {
+		case "":
+			sr = migrate.RealmConn(c.Driver, nil)
+		default:
+			sr = migrate.SchemaConn(c.Driver, c.URL.Schema, nil)
+		}
+		return &stateReadCloser{
+			StateReader: sr,
+			Closer:      c,
+			Schema:      c.URL.Schema,
+		}, nil
+	}
+}
+
+// hclStateReadr returns a migrate.StateReader that reads the state from the given HCL paths urls.
+func hclStateReader(ctx context.Context, config *stateReaderConfig, urls []*url.URL) (*stateReadCloser, error) {
+	paths := make([]string, len(urls))
+	for i, u := range urls {
+		paths[i] = u.Path
+	}
+	parser, err := parseHCLPaths(paths...)
+	if err != nil {
+		return nil, err
+	}
+	realm := &schema.Realm{}
+	if err := config.dev.Eval(parser, realm, nil); err != nil {
+		return nil, err
+	}
+	if len(config.schemas) > 0 {
+		// Validate all schemas in file were selected by user.
+		sm := make(map[string]bool, len(config.schemas))
+		for _, s := range config.schemas {
+			sm[s] = true
+		}
+		for _, s := range realm.Schemas {
+			if !sm[s.Name] {
+				return nil, fmt.Errorf("schema %q from paths %q is not requested (all schemas in HCL must be requested)", s.Name, paths)
+			}
+		}
+	}
+	// In case the dev connection is bound to a specific schema, we require the
+	// desired schema to contain only one schema. Thus, executing diff will be
+	// done on the content of these two schema and not the whole realm.
+	if config.dev.URL.Schema != "" && len(realm.Schemas) > 1 {
+		return nil, fmt.Errorf(
+			"cannot use HCL with more than 1 schema when dev-url is limited to schema %q",
+			config.dev.URL.Schema,
+		)
+	}
+	if norm, ok := config.dev.Driver.(schema.Normalizer); ok && len(realm.Schemas) > 0 {
+		realm, err = norm.NormalizeRealm(ctx, realm)
+		if err != nil {
+			return nil, err
+		}
+	}
+	t := &stateReadCloser{StateReader: migrate.Realm(realm)}
+	if len(realm.Schemas) == 1 {
+		t.Schema = realm.Schemas[0].Name
+	}
+	return t, nil
+}
+
+// Close redirects calls to Close to the enclosed io.Closer.
+func (sr *stateReadCloser) Close() {
+	if sr.Closer != nil {
+		sr.Closer.Close()
+	}
 }
